slas-2024-lint fixes
